services/preprocessor/pyproject.toml
[project]
name = "video-verification-preprocessor"
version = "1.0.0"
description = "Preprocessor service: video → frame extraction"
requires-python = ">=3.11"
dependencies = [
    "video-verification-shared",
    "confluent-kafka>=2.3",
    "av>=12.0",
    "Pillow>=10.0",
    "structlog>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["preprocessor*"]
